requirements.txt
uvicorn[standard]
fastapi
sqlalchemy
pymysql
pydantic
pydantic[email]
passlib
requests
PyJWT
